# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminUpdateUserRequest:
    properties:
      username:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
...
  dto.PublicUserResponse:
    properties:
      avatar:
        type: string
      avatarCard:
        type: string
      avatarLqip:
        type: string
      avatarThumb:
        type: string
      completeName:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      language:
        type: string
      username:
        type: string
    type: object
  dto.UpdateRoleRequest:
    properties:
      role:
        type: string
    type: object
  dto.UserResponse:
    properties:
      avatar:
        type: string
      avatarCard:
        type: string
      avatarLqip:
        type: string
      avatarStatus:
        type: string
      avatarThumb:
        type: string
      completeName:
        type: string
      email:
        type: string
      id:
        type: string
      language:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Backend API for JojoOnePieceSimulator2 - Google-only auth, Stands and
    Devil Fruits catalogues.
  title: JojoOnePieceSimulator2 API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Admin only.
      parameters:
      - description: Page size (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/{id}:
    delete:
      description: Admin only. An admin cannot delete their own account through this
        route, and the last remaining admin cannot be deleted.
      parameters:
      - description: User id (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete another user
      tags:
      - users
    get:
      description: Never includes email or role.
      parameters:
      - description: User id (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PublicUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get another user's public profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: User id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New username
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change another user's username
      tags:
      - users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Admin only. An admin cannot change their own role through this
        route (use another admin account), and the last remaining admin cannot be
        demoted. A role change only takes effect on the affected user's next login,
        since the role is carried in the JWT. ADMIN_EMAILS re-promotes a listed email
        on every login, so demoting one of those emails is temporary.
      parameters:
      - description: User id (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change another user's role
      tags:
      - users
  /users/me:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete the authenticated caller's account
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the authenticated caller's profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: The only self-editable field. email, role, and completeName are
        never accepted here - sending them is a 400.
      parameters:
      - description: New username
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update the authenticated caller's username
      tags:
      - users
  /users/me/picture:
    delete:
      description: Reverts display back to the Google-synced picture.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete the authenticated caller's avatar
      tags:
      - users
    patch:
      consumes:
      - multipart/form-data
      description: Uploads the image to object storage; the response's avatar fields
        are the previous renditions until the background worker finishes.
      parameters:
      - description: Image file (WebP, AVIF, JPEG, PNG or GIF)
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Set the authenticated caller's avatar
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"